Session Management
Implemented basic session management and revised some of the Entity
Framework relations.  Additional table in the database stores
information about every user's session.  This will be used to
authenticate certain WCF methods without the need for username and
password as well as remembering the logged in user for a longer period
of time.