Session Management + Code Refactor
Implemented first basic version of the session management which provides
the ability for users to log in automatically using the session
management token from their android device.

Also refactored and provided headers and additional comments for all of
the classes within the solution.  Now beginning to work on the carshare
service.