UI, Search Service
Improved android client's UI and implemented fragments.
The client now shows a list of car shares that the currently logged in
user either participates in or is the driver of.
Right now, I have also started implementing the web service for
searching car shares.
Modified ServiceHostFactory to retrieve the right type of service based
on request.