Search fragment, handling screen orientation.
Currently working on the search animation, expanding and shrinking of
the search panel and search results listview.

Cleaned up a lot of old experimental code.
Also working on handling screen orientation changes, saving and
retrieving activity and fragment states. Still getting a number of
nullreference exceptions, must come up with a universal way of handling
these. I don't really want to disable screen orientation changes in the
app as it shows bad design.

Also made a few changes to the way data is passed between activity and
fragments.  Data is now serialized as json string and passed in a
bundle.