Offering a journey
Beginning to work on the activity allowing users to create a new
journey.
Also removed the concept of a journey having a departure and destination
points. Each journey now has a collection of locations which will enable
users to specify start and end locations together with waypoints through
which the journey will take place.